Add functions for calculating and setting the kl divergence to EmceePTFile
EmceePT calculates the evidence when finalizing and makes that available in the file.
This is crucial for calculating the kl divergence.
For this reason, and because regular Emcee does not currently support calculation of the evidence,
kl divergence is only implemented in EmceePT.
The EmceePT sampler now also sets the kl divergence in the file when calling finalize().