Using inverse waveform transforms instead for getting the injection probabilities.
Normally, the waveform transforms are responsible for transforming the set of parameters that the model is given in (the "variable parameters") to the set of parameters that the waveform generator understands.
In the case of checking the likelihood at the injection value, we already have the parameters that are understandable by the waveform generator (from the injection file), but we need the variable parameters to check if they are inside the prior region and get their prior values.
So we want to apply the inverse of the waveform transforms.